+++ lookups/CH_lookup.py
@@ -3,8 +3,9 @@
 
 import time
 import feedparser
-import urllib
-from .base_lookup import *
+from urllib.parse import urlencode
+from common import *
+from . base_lookup import *
 
 # ============================================================================
 class CH_lookup(base_lookup):
@@ -29,47 +30,45 @@ class CH_lookup(base_lookup):
 
         # if we do have an APIKey, then use it (create a dict to merge), if we
         # don't have one, use an empty dict instead
-        if self.C.APIKey:
-            di_APIKey = {"key": self.C.APIKey}
+        if C.APIKey:
+            di_APIKey = {"key": C.APIKey}
         else:
             di_APIKey = {}
-# todo: need to replace self.C.APIurl (which is currently coming from
+# todo: need to replace C.APIurl (which is currently coming from
 #       common/settings/constants.py
-        rss = feedparser.parse("%s?%s" % (self.C.APIurl, urllib.urlencode(
-                          dict({"was": question, "maxnum":1}, **di_APIKey)
+        rss = feedparser.parse("%s?%s" % (C.APIurl, urlencode(
+                          { **{"was": question, "maxnum": 1}, **di_APIKey }
               )))
-        # python3: { **{"was": question, "maxnum": 1}, **{"key": ST_APIKey} }
-        # python3: { **{"was": question, "maxnum": 1}, **di_APIKey }
 
         # lookup unsuccessful
         if (rss.status != 200) or (len(rss.entries) < 1):
             if rss.status != 200:
-                self.L.log(severity="W", msg='ID=%s status=%s msg="lookup unsuccessful"' % (
+                L.log(severity="W", msg='ID=%s status=%s msg="lookup unsuccessful"' % (
                                     self.ID, rss.status))
             elif len(rss.entries) < 1:
-                self.L.log(severity="D", msg='action=do_lookup ID=%s msg="no results returned"' % (
+                L.log(severity="D", msg='action=do_lookup ID=%s msg="no results returned"' % (
                                     self.ID))
             if question in cache:
-                self.L.log(severity="I", msg='ID="%s" location=%s answer="%s"' % (
+                L.log(severity="I", msg='ID="%s" location=%s answer="%s"' % (
                       self.ID, cache[question]["cache_type"] + "_expired",
                       cache[question]["title"]))
                 return (cache[question]["title"])
             else:
-                self.L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
+                L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
                       self.ID, "lookup_failed", question))
                 return (question)
 
         # lookup successful
-        if self.C.DEBUG:
+        if C.DEBUG:
             entrynum=0
             for entry in rss.entries:
                 entrynum += 1
-                for key in entry.keys():
-                    self.L.log(severity="D", msg="action=do_lookup ID=%s entry=%d %s=%s" % (
+                for key in list(entry.keys()):
+                    L.log(severity="D", msg="action=do_lookup ID=%s entry=%d %s=%s" % (
                           self.ID, entrynum, key, entry[key]))
 
         try:
-            self.L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
+            L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
                       self.ID, "lookup_succeeded", rss.entries[0].title))
             cache[question] = {"title": rss.entries[0].title,
                                "date_last_update": int(time.time()),
@@ -77,15 +76,14 @@ class CH_lookup(base_lookup):
                               }
             return (rss.entries[0].title)
         except:
-            self.L.log(severity="W",
+            L.log(severity="W",
                   msg='ID=%s msg="lookup successful but result not parsable"'\
                       % self.ID)
-            self.L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
+            L.log(severity="I", msg='ID=%s location=%s answer="%s"' % (
                       self.ID, "lookup_failed", question))
             cache[question] = {"title": question,
                                "date_last_update": int(time.time()),
                                "cache_type": "negative",
-                               "cache_version": self.C.cache_version,
                               }
         # fallback
         return (question)