+++ lookups/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: UTF-8 -*-
 # vim: set et ai ci sm tw=78 si sw=4 ru filetype=python fileencoding=utf-8 :
 
-
-from .base_lookup import *
-from .CH_lookup import *
+from . base_lookup import *
+from . CH_lookup import *